moving more classes to their respective files
