+++ src/Makefile.am
@@ -8,7 +8,9 @@ njnrSOURCES += CodeGenerator.cpp common/symbol_table.cpp common/symbol_table_sta
 njnrSOURCES += common/symbol_table_stackX.cpp common/symbol_table_entry.cpp
 njnrSOURCES += expressions/expressions.cpp
 njnrSOURCES += functions/function_calls.cpp functions/functions.cpp
-njnrSOURCES += statements/statements.cpp
+njnrSOURCES += statements/statements.cpp Constant.cpp FunctionBinding.cpp
+njnrSOURCES += Identifier.cpp ParameterBinding.cpp ReturnPacket.cpp Statement.cpp
+njnrSOURCES += TranslationUnit.cpp VariableBinding.cpp
 njnrSOURCES += include/debug.hpp include/type.hpp include/List.hpp
 njnrSOURCES += include/ListNode.hpp include/Constant.hpp include/common/symbol_table_stackX.hpp
 njnrSOURCES += include/common/symbol_table_stack.hpp include/common/symbol_table.hpp
