+++ src/VariableBinding.cpp
@@ -1,0 +1,17 @@
+#include <string>
+
+#include "VariableBinding.hpp"
+#include "type.hpp"
+#include "Identifier.hpp"
+using namespace njnr;
+
+   Varb::Varb() : Identifier{} {}
+   Varb::~Varb() {
+                  report(njnr::logType::debug, "running Varb() Destructor");
+                  report(njnr::logType::debug, this->toString());
+                 }
+
+   const std::string Varb::toString() const
+   {
+      return ReturnPacket::toString();
+   }
