+++ src/cpptypes.cpp
@@ -1,424 +1,12 @@
 #include <config.h>
+#include <string>
+
 #include "type.hpp"
 #include "List.hpp"
-#include "ReturnPacket.hpp"
-#include "TranslationUnit.hpp"
 #include "ParameterBinding.hpp"
-#include "symbol_table_entry.hpp"
-#include "Compiler.hpp"
 
 namespace njnr
 {
-   ReturnPacket::ReturnPacket() : m_pair{},
-                                  funcent{nullptr},
-                                  params{0},
-                                  offset{0},
-                                  lval{false},
-                                  numeric{false},
-                                  ttype{njnr::type::INT}
-   {}
-
-   ReturnPacket::ReturnPacket(bool lval,
-                              njnr::type ttype = njnr::type::INT,
-                              bool ifnum = false,
-                              int inoffset = 0) :
-                                                 m_pair{},
-                                                 funcent{nullptr},
-                                                 params{0},
-                                                 offset{inoffset},
-                                                 lval{lval},
-                                                 numeric{ifnum},
-                                                 ttype{ttype}
-   {}
-
-   const bool ReturnPacket::getlval() const
-   {
-       return lval;
-   }
-   void ReturnPacket::setlval(const bool in)
-   {
-       lval = in;
-   }
-   const njnr::type  ReturnPacket::gettype() const
-   {
-       return ttype;
-   }
-   void ReturnPacket::settype(const njnr::type in)
-   {
-       ttype = in;
-   }
-   const bool ReturnPacket::getnumeric() const {
-       return numeric;
-   }
-   void ReturnPacket::setnumeric(const bool in)
-   {
-       numeric = in;
-   }
-   int ReturnPacket::getoffset() const
-   {
-       return offset;
-   }
-   void ReturnPacket::setoffset(const int in)
-   {
-       offset = in;
-   }
-
-   const std::string ReturnPacket::toString() const
-   {
-    /**
-         int offset;
-         bool lval;
-         bool numeric;
-         njnr::type ttype;
-
-         struct Pair m_pair;
-         S_TableEntry* funcent;
-         int params;
-     */
-      std::string r{"lval: "};
-      r += (lval == true ? "true" : "false");
-      r += "\noffset: ";
-      r += std::to_string(offset);
-      r += "\nnumeric: ";
-      r += (numeric == true ? "true" : "false");
-      r += "\nparameters: ";
-      r += std::to_string(params);
-      r += "\nttype: ";
-      r += std::to_string(static_cast<int>(ttype));
-
-      r += "\n" + m_pair.toString() + "\n";
-      if (nullptr != funcent)
-      {
-        r += static_pointer_cast<S_TableEntryX>(funcent)->toString() + "\n";
-      }
-      return r;
-   };
-
-
-   Constant::Constant() : ReturnPacket{},
-                          val{""},
-                          typ{njnr::type::VOID}
-                          {}
-   Constant::Constant(bool lval, njnr::type ttype, bool ifnum, int offset) :
-                      ReturnPacket{lval, ttype, ifnum, offset},
-                      val{""},
-                      typ{njnr::type::VOID}
-                      {}
-   Constant::Constant(std::string val, njnr::type t): val{val}, typ{t} {}
-
-   std::string Constant::getValue() const
-   {
-      return val;
-   }
-
-   void Constant::setValue(const std::string in)
-   {
-      val = in;
-   }
-
-   njnr::type Constant::getType() const
-   {
-      return typ;
-   }
-
-   void Constant::setType(const njnr::type t)
-   {
-      typ = t;
-   }
-
-
-   const std::string Identifier::toString() const
-   {
-      std::string ret = ReturnPacket::toString();
-      return "\nvalue: " + value + "\n" + ret;
-   }
-   Identifier::Identifier() : Constant{false, njnr::type::IDENT, false, 0},
-                              value{""} {}
-
-   Identifier::Identifier(const std::string invalue) :
-                          Constant{false, njnr::type::IDENT, false, 0},
-                          value{invalue}
-                          {}
-
-   Identifier::Identifier(const Constant in) :
-                          Constant{in.getlval(),
-                          njnr::type::IDENT, false, in.getoffset()},
-                          value{in.getValue()}
-                          {}
-
-   Identifier& Identifier::operator=(const Identifier& in)
-   {
-       if (&in != this)
-       {
-           value = in.value;
-           lval = in.lval;
-           numeric = false;
-           ttype = njnr::type::IDENT;
-           offset = in.offset;
-       }
-       return *this;
-   }
-   std::string Identifier::getvalue() const
-   {
-       return value;
-   }
-
-   void Identifier::setvalue(const std::string in)
-   {
-       value = in;
-   }
-
-   Funcb::Funcb() : param_type{},
-       returntype{njnr::type::VOID},
-       bodydef{false},
-       num_param{0},
-       label{0},
-       localcount{0},
-       actual_num{0},
-       funcheader{nullptr},
-       funcbody_list{nullptr}
-   {}
-   Funcb::Funcb(njnr::type returntype, bool bodydef, int num_param,
-                std::vector<njnr::type> param_type, int label,
-                int localcount, int actual_num) :
-       param_type{param_type},
-       returntype{returntype},
-       bodydef{bodydef},
-       num_param{num_param},
-       label{label},
-       localcount{localcount},
-       actual_num{actual_num},
-       funcheader{nullptr},
-       funcbody_list{nullptr}
-   {}
-   Funcb::Funcb(njnr::type returntype) : Funcb{}
-   {
-       this->returntype = returntype;
-   }
-   Funcb::Funcb(const Funcb& in) :
-       param_type{in.param_type},
-       returntype{in.returntype},
-       bodydef{in.bodydef},
-       num_param{in.num_param},
-       label{in.label},
-       localcount{in.localcount},
-       actual_num{in.actual_num},
-       funcheader{in.funcheader},
-       funcbody_list{in.funcbody_list}
-   {
-       setvalue(in.getvalue());
-   }
-
-   Funcb& Funcb::operator=(const Funcb& in)
-   {
-       if (&in != this) {
-           returntype = in.returntype;
-           bodydef = in.bodydef;
-           num_param = in.num_param;
-           label = in.label;
-           localcount = in.localcount;
-           actual_num = in.actual_num;
-           param_type = in.param_type;
-           funcheader = in.funcheader;
-           funcbody_list = in.funcbody_list;
-           setvalue(in.getvalue());
-       }
-       return *this;
-   }
-   Funcb::~Funcb()
-   {
-      report(njnr::logType::debug, "running Funcb() destructor");
-      report(njnr::logType::debug, "Funcb:" + this->toString());
-/*
-      report(njnr::logType::debug,
-             "bodydef"+ bodydef);
-      report(njnr::logType::debug,
-             "num_param"+ num_param);
-      report(njnr::logType::debug,
-             "label"+ label);
-      report(njnr::logType::debug,
-             "localcount"+ localcount);
-      report(njnr::logType::debug,
-             "actual_num"+ actual_num);
-*/
-      param_type.clear();
-      funcbody_list = nullptr;
-      funcheader = nullptr;
-    }
-   std::vector<njnr::type> Funcb::getparam_type()
-   {
-       return param_type;
-   }
-   njnr::type Funcb::getreturntype()
-   {
-       return returntype;
-   }
-   bool Funcb::getbodydef()
-   {
-       return bodydef;
-   }
-   int Funcb::getnum_param()
-   {
-       return num_param;
-   }
-   int Funcb::getlabel()
-   {
-       return label;
-   }
-   int Funcb::getlocalcount()
-   {
-       return localcount;
-   }
-   int Funcb::getactual_num()
-   {
-       return actual_num;
-   }
-   void Funcb::Funcb::setparam_type(std::vector<njnr::type> param_type)
-   {
-       this->param_type = param_type;
-   }
-   void Funcb::setreturntype(njnr::type returntype)
-   {
-       this->returntype = returntype;
-   }
-   void Funcb::setbodydef(bool bodydef)
-   {
-       this->bodydef = bodydef;
-   }
-   void Funcb::setnum_param(int num_param)
-   {
-       this->num_param = num_param;
-   }
-   void Funcb::setlabel(int label)
-   {
-       this->label = label;
-   }
-   void Funcb::setlocalcount(int localcount)
-   {
-       this->localcount = localcount;
-   }
-   void Funcb::setactual_num(int actual_num )
-   {
-       this->actual_num = actual_num;
-   }
-
-   void Funcb::setfuncbody_list(std::shared_ptr<List> funcbody)
-   {
-      funcbody_list = funcbody;
-   }
-   std::shared_ptr<List> Funcb::getfuncbody_list(void)
-   {
-    return funcbody_list;
-   }
-   void Funcb::setfuncheader(std::shared_ptr<funcheadertype> funcheader)
-   {
-      this->funcheader = funcheader;
-   }
-   std::shared_ptr<funcheadertype> Funcb::getfuncheader(void)
-   {
-    return funcheader;
-   }
-   const std::string Funcb::toString() const
-   {
-    /**
-         std::vector<njnr::type> param_type;
-         type   returntype;
-         bool   bodydef;
-         int    num_param;
-         int    label;
-         int    localcount;
-         int    actual_num;
-         std::shared_ptr<funcheadertype> funcheader;
-         std::shared_ptr<List> funcbody_list;
-     */
-
-      std::string r{};
-      r += "Currently Set Return Type: ";
-      r += Compiler::getStringFromType(returntype);
-      r += "bodydef:? " + (true == bodydef) ? "true\n" : "false\n";
-      r += "label: " + std::to_string(label) + "\n";
-      r += "localcount: " + std::to_string(localcount) + "\n";
-      r += "actual_num" + std::to_string(actual_num) + "\n";
-      r += "num_param" + std::to_string(num_param) + "\n";
-      if (0 <= param_type.size())
-      {
-        for (auto t : param_type)
-        {
-            r += "param type: " + njnr::typeToStringMap.at(t) + "\n"; 
-        }
-      }
-/* TODO : params list vector */
-      if (nullptr != funcheader)
-      {
-         r += funcheader->toString() + "\n";
-      }
-      if (nullptr != funcbody_list)
-      {
-         r += funcbody_list->toString();
-      }
-      return r;
-   }
-
-   Varb::Varb() : Identifier{} {}
-   Varb::~Varb() {
-                  report(njnr::logType::debug, "running Varb() Destructor");
-                  report(njnr::logType::debug, this->toString());
-                 }
-
-   const std::string Varb::toString() const
-   {
-      return ReturnPacket::toString();
-   }
-
-   const std::string Paramb::toString() const
-   {
-      return ReturnPacket::toString();
-   }
-   Paramb::Paramb() : ReturnPacket{} {}
-   Paramb::~Paramb()
-   {
-      report(njnr::logType::debug,
-             "running Paramb() Destructor");
-      report(njnr::logType::debug, this->toString());
-
-   }
-
-   const std::string Translation_Unit::toString() const
-   {
-      std::string ret{ReturnPacket::toString()};
-//         std::shared_ptr<ReturnPacket> translation{nullptr};
-//         trans_unit_type trans_type;
-
-      return "translation unit: " + this->toString() + "\n" + ret;
-   }
-   Translation_Unit::Translation_Unit() : translation{nullptr},
-                                          trans_type{trans_unit_type::INVALID}
-                                          {}
-
-   Translation_Unit::~Translation_Unit()
-   {
-      report(njnr::logType::debug,
-             "running Translation_Unit() Destructor");
-      report(njnr::logType::debug, this->toString());
-   }
-
-   std::shared_ptr<ReturnPacket> Translation_Unit::get_translation()
-   {
-      return translation;
-   }
-   void Translation_Unit::set_translation(std::shared_ptr<ReturnPacket> translation)
-   {
-      this->translation = translation;
-   }
-   void Translation_Unit::set_trans_unit_type(trans_unit_type intype)
-   {
-      trans_type = intype;
-   }
-   trans_unit_type Translation_Unit::get_trans_unit_type()
-   {
-      return trans_type;
-   }
-
    const std::string funcheadertype::toString() const
    {
       std::string r = "";
